quiz_file/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
flask-limiter==3.5.0
werkzeug==3.0.1